requirements.txt
torch==2.5.1
torchvision==0.20.1
numpy==1.26.4
pillow==11.3.0
opencv-python
plyfile
huggingface_hub==0.35.0
safetensors
hydra-core
pandas==2.3.2
easydict
prettytable
accelerate==1.10.1
rich
scipy
h5py
transformers==4.56.2
tqdm